[IMP] Improvements to the entry point.
- Support for commands using argparse to avoid any external dependency, only standard library
- Keep compatibility with the current behaviour
- Add support to update the database using click-odoo-update from click-odoo-contrib

These changes will help us to speed up the database because click-odoo-update only updates the modules that changed since last update, the worst time is when all the modules have changed, in such case is the same as the current behaviour. The run command was added because we detected that not all parameters were added to the config file and all the env vars have to be pased before any update or execution.